# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "get": {
        "operationId": "UsersController_getAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users returned for the current actor scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "List users visible to the acting role",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Only System Admin can create users."
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "U201",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "U201",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deactivated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Deactivate a user",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/returns/history": {
      "get": {
        "operationId": "ReturnsController_getHistory",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return history returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReturnHistoryResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Get return history visible to the acting role",
        "tags": [
          "returns"
        ]
      }
    },
    "/api/returns/{resourceId}/process": {
      "post": {
        "operationId": "ReturnsController_process",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Acting role for RBAC. Allowed values: Requestor, Dept Head, Registrar, Staff, System Admin.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Staff"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Acting user id used for department scoping and own-record views. Example: U100.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "U100"
            }
          },
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "RES-ITL-001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnProcessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return processed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid workflow transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Role or department scope does not allow this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Requested entity was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "role-header": []
          },
          {
            "user-id-header": []
          }
        ],
        "summary": "Process a returned resource",
        "tags": [
          "returns"
        ]
      }
    },
...
  },
  "info": {
    "title": "ResourceX Demo Backend",
    "description": "Demo-exact NestJS backend for the ResourceX frontend. All authorization is role-based through request headers. Use x-user-role on every protected route and x-user-id when actor-specific scoping is required.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Required RBAC header. Example values: Requestor, Dept Head, Registrar, Staff, System Admin."
      },
      "user-id-header": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Optional acting user id header for own-profile and department-scoped behavior. Example: U100."
      }
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "example": "Selected resources must match request quantity."
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "UserPreferencesDto": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "notifications"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "U1"
          },
          "name": {
            "type": "string",
            "example": "RAVI CHANDRA"
          },
          "email": {
            "type": "string",
            "example": "ravi@resourcex.com"
          },
          "password": {
            "type": "string",
            "example": "12345678"
          },
          "role": {
            "type": "string",
            "example": "Requestor"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "preferences": {
            "$ref": "#/components/schemas/UserPreferencesDto"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "U201"
          },
          "name": {
            "type": "string",
            "example": "Asha Nair"
          },
          "email": {
            "type": "string",
            "example": "asha@resourcex.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "Requestor",
              "Dept Head",
              "Registrar",
              "Staff",
              "System Admin"
            ],
            "example": "Staff"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ],
            "example": "Active"
          },
          "password": {
            "type": "string",
            "example": "temporary123"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "U201"
          },
          "name": {
            "type": "string",
            "example": "Asha Nair"
          },
          "email": {
            "type": "string",
            "example": "asha@resourcex.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "Requestor",
              "Dept Head",
              "Registrar",
              "Staff",
              "System Admin"
            ],
            "example": "Staff"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "Inactive",
              "Suspended"
            ],
            "example": "Active"
          },
          "password": {
            "type": "string",
            "example": "newsecret123"
          }
        }
      },
...
      "ResourceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "RES-1049"
          },
          "name": {
            "type": "string",
            "example": "Laptop - Dell XPS"
          },
          "type": {
            "type": "string",
            "example": "Laptop"
          },
          "department": {
            "type": "string",
            "example": "IT Services"
          },
          "serialNumber": {
            "type": "string",
            "example": "SN-998822"
          },
          "status": {
            "type": "string",
            "example": "Allocated"
          },
          "condition": {
            "type": "string",
            "example": "Good"
          },
          "assignedTo": {
            "type": "string",
            "example": "RAVI CHANDRA"
          },
          "assignedToId": {
            "type": "string",
            "example": "U1"
          },
          "date": {
            "type": "string",
            "example": "Jan 12, 2023"
          },
          "vendor": {
            "type": "string",
            "example": "Cisco Systems"
          },
          "invoice": {
            "type": "string",
            "example": "INV-8899"
          },
          "location": {
            "type": "string",
            "example": "IT Lab"
          },
          "returnRequestedBy": {
            "type": "string",
            "example": "RAVI CHANDRA"
          },
          "returnRequestedById": {
            "type": "string",
            "example": "U1"
          }
        },
        "required": [
          "id",
          "type",
          "department",
          "status"
        ]
      },
...
      "ReturnHistoryResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "RES-9988"
          },
          "type": {
            "type": "string",
            "example": "Laptop - Old"
          },
          "returnedBy": {
            "type": "string",
            "example": "Bob Builder"
          },
          "returnDate": {
            "type": "string",
            "example": "Oct 20, 2023"
          },
          "processDate": {
            "type": "string",
            "example": "Oct 21, 2023"
          },
          "condition": {
            "type": "string",
            "example": "Bad"
          },
          "finalStatus": {
            "type": "string",
            "example": "Scrapped"
          },
          "department": {
            "type": "string",
            "example": "Operations"
          }
        },
        "required": [
          "code",
          "type",
          "returnedBy",
          "returnDate",
          "processDate",
          "condition",
          "finalStatus",
          "department"
        ]
      },
      "ReturnProcessDto": {
        "type": "object",
        "properties": {
          "condition": {
            "type": "string",
            "enum": [
              "New",
              "Good",
              "Average",
              "Fair",
              "Damaged",
              "Bad"
            ],
            "example": "Good"
          }
        },
        "required": [
          "condition"
        ]
      },
...
    }
  }
}